requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fuzzywuzzy==0.18.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
lxml==4.8.0
Markdown==3.3.6
MouseInfo==0.1.3
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc==8.4.1
pyobjc-core==8.4.1
pyobjc-framework-Accessibility==8.4.1
pyobjc-framework-Accounts==8.4.1
pyobjc-framework-AddressBook==8.4.1
pyobjc-framework-AdServices==8.4.1
pyobjc-framework-AdSupport==8.4.1
pyobjc-framework-AppleScriptKit==8.4.1
pyobjc-framework-AppleScriptObjC==8.4.1
pyobjc-framework-ApplicationServices==8.4.1
pyobjc-framework-AppTrackingTransparency==8.4.1
pyobjc-framework-AudioVideoBridging==8.4.1
pyobjc-framework-AuthenticationServices==8.4.1
pyobjc-framework-AutomaticAssessmentConfiguration==8.4.1
pyobjc-framework-Automator==8.4.1
pyobjc-framework-AVFoundation==8.4.1
pyobjc-framework-AVKit==8.4.1
pyobjc-framework-BusinessChat==8.4.1
pyobjc-framework-CalendarStore==8.4.1
pyobjc-framework-CallKit==8.4.1
pyobjc-framework-CFNetwork==8.4.1
pyobjc-framework-ClassKit==8.4.1
pyobjc-framework-CloudKit==8.4.1
pyobjc-framework-Cocoa==8.4.1
pyobjc-framework-Collaboration==8.4.1
pyobjc-framework-ColorSync==8.4.1
pyobjc-framework-Contacts==8.4.1
pyobjc-framework-ContactsUI==8.4.1
pyobjc-framework-CoreAudio==8.4.1
pyobjc-framework-CoreAudioKit==8.4.1
pyobjc-framework-CoreBluetooth==8.4.1
pyobjc-framework-CoreData==8.4.1
pyobjc-framework-CoreHaptics==8.4.1
pyobjc-framework-CoreLocation==8.4.1
pyobjc-framework-CoreMedia==8.4.1
pyobjc-framework-CoreMediaIO==8.4.1
pyobjc-framework-CoreMIDI==8.4.1
pyobjc-framework-CoreML==8.4.1
pyobjc-framework-CoreMotion==8.4.1
pyobjc-framework-CoreServices==8.4.1
pyobjc-framework-CoreSpotlight==8.4.1
pyobjc-framework-CoreText==8.4.1
pyobjc-framework-CoreWLAN==8.4.1
pyobjc-framework-CryptoTokenKit==8.4.1
pyobjc-framework-DataDetection==8.4.1
pyobjc-framework-DeviceCheck==8.4.1
pyobjc-framework-DictionaryServices==8.4.1
pyobjc-framework-DiscRecording==8.4.1
pyobjc-framework-DiscRecordingUI==8.4.1
pyobjc-framework-DiskArbitration==8.4.1
pyobjc-framework-DVDPlayback==8.4.1
pyobjc-framework-EventKit==8.4.1
pyobjc-framework-ExceptionHandling==8.4.1
pyobjc-framework-ExecutionPolicy==8.4.1
pyobjc-framework-ExternalAccessory==8.4.1
pyobjc-framework-FileProvider==8.4.1
pyobjc-framework-FileProviderUI==8.4.1
pyobjc-framework-FinderSync==8.4.1
pyobjc-framework-FSEvents==8.4.1
pyobjc-framework-GameCenter==8.4.1
pyobjc-framework-GameController==8.4.1
pyobjc-framework-GameKit==8.4.1
pyobjc-framework-GameplayKit==8.4.1
pyobjc-framework-ImageCaptureCore==8.4.1
pyobjc-framework-IMServicePlugIn==8.4.1
pyobjc-framework-InputMethodKit==8.4.1
pyobjc-framework-InstallerPlugins==8.4.1
pyobjc-framework-InstantMessage==8.4.1
pyobjc-framework-Intents==8.4.1
pyobjc-framework-IntentsUI==8.4.1
pyobjc-framework-IOSurface==8.4.1
pyobjc-framework-iTunesLibrary==8.4.1
pyobjc-framework-KernelManagement==8.4.1
pyobjc-framework-LatentSemanticMapping==8.4.1
pyobjc-framework-LaunchServices==8.4.1
pyobjc-framework-libdispatch==8.4.1
pyobjc-framework-LinkPresentation==8.4.1
pyobjc-framework-LocalAuthentication==8.4.1
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.4.1
pyobjc-framework-MailKit==8.4.1
pyobjc-framework-MapKit==8.4.1
pyobjc-framework-MediaAccessibility==8.4.1
pyobjc-framework-MediaLibrary==8.4.1
pyobjc-framework-MediaPlayer==8.4.1
pyobjc-framework-MediaToolbox==8.4.1
pyobjc-framework-Metal==8.4.1
pyobjc-framework-MetalKit==8.4.1
pyobjc-framework-MetalPerformanceShaders==8.4.1
pyobjc-framework-MetalPerformanceShadersGraph==8.4.1
pyobjc-framework-MetricKit==8.4.1
pyobjc-framework-MLCompute==8.4.1
pyobjc-framework-ModelIO==8.4.1
pyobjc-framework-MultipeerConnectivity==8.4.1
pyobjc-framework-NaturalLanguage==8.4.1
pyobjc-framework-NetFS==8.4.1
pyobjc-framework-Network==8.4.1
pyobjc-framework-NetworkExtension==8.4.1
pyobjc-framework-NotificationCenter==8.4.1
pyobjc-framework-OpenDirectory==8.4.1
pyobjc-framework-OSAKit==8.4.1
pyobjc-framework-OSLog==8.4.1
pyobjc-framework-PassKit==8.4.1
pyobjc-framework-PencilKit==8.4.1
pyobjc-framework-Photos==8.4.1
pyobjc-framework-PhotosUI==8.4.1
pyobjc-framework-PreferencePanes==8.4.1
pyobjc-framework-PushKit==8.4.1
pyobjc-framework-Quartz==8.4.1
pyobjc-framework-QuickLookThumbnailing==8.4.1
pyobjc-framework-ReplayKit==8.4.1
pyobjc-framework-SafariServices==8.4.1
pyobjc-framework-SceneKit==8.4.1
pyobjc-framework-ScreenSaver==8.4.1
pyobjc-framework-ScreenTime==8.4.1
pyobjc-framework-ScriptingBridge==8.4.1
pyobjc-framework-SearchKit==8.4.1
pyobjc-framework-Security==8.4.1
pyobjc-framework-SecurityFoundation==8.4.1
pyobjc-framework-SecurityInterface==8.4.1
pyobjc-framework-ServiceManagement==8.4.1
pyobjc-framework-ShazamKit==8.4.1
pyobjc-framework-Social==8.4.1
pyobjc-framework-SoundAnalysis==8.4.1
pyobjc-framework-Speech==8.4.1
pyobjc-framework-SpriteKit==8.4.1
pyobjc-framework-StoreKit==8.4.1
pyobjc-framework-SyncServices==8.4.1
pyobjc-framework-SystemConfiguration==8.4.1
pyobjc-framework-SystemExtensions==8.4.1
pyobjc-framework-UniformTypeIdentifiers==8.4.1
pyobjc-framework-UserNotifications==8.4.1
pyobjc-framework-UserNotificationsUI==8.4.1
pyobjc-framework-VideoSubscriberAccount==8.4.1
pyobjc-framework-VideoToolbox==8.4.1
pyobjc-framework-Virtualization==8.4.1
pyobjc-framework-Vision==8.4.1
pyobjc-framework-WebKit==8.4.1
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.1.1
python-docx==0.8.11
python-Levenshtein==0.12.2
pytweening==1.0.4
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
rubicon-objc==0.4.2
six==1.16.0
soupsieve==2.3.1
teacherhelper==1.0.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
